ME_QCALC_MODULE_INHERITANCE

+++ src/main/java/com/crio/qcalc/StandardCalculator.java
@@ -16,7 +16,7 @@ public class StandardCalculator {
     
     }    
 
-    public int getResult() {
+    public double getResult() {
 
         return result;
